requirements.txt
args==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
clint==0.5.1
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
oauthlib==3.1.0
packaging==20.9
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
pyparsing==2.4.7
PySocks==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
tweepy==3.8.0
uritemplate==3.0.1
urllib3==1.25.9
